# next.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inbound API v2",
    "version": "2.0.0",
    "description": "The official API documentation for Inbound - Email API Service. Manage domains, send emails, handle webhooks, and more.",
    "contact": {
      "name": "Inbound Support",
      "url": "https://inbound.new/support",
      "email": "support@inbound.new"
    },
    "license": {
      "name": "Commercial",
      "url": "https://inbound.new/terms"
    }
  },
  "servers": [
    {
      "url": "https://inbound.new/api/v2",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api/v2",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "API Key authentication. Use 'Bearer YOUR_API_KEY'"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for session-based authentication"
      }
    },
...
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Authentication required. Please provide a valid API key or session token."
                },
                "code": {
                  "type": "string",
                  "example": "401"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred. Please try again later."
                },
                "code": {
                  "type": "string",
                  "example": "500"
                },
                "details": {
                  "type": "string",
                  "example": "Additional error details"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/v2/mail": {
      "get": {
        "operationId": "get-v2-mail",
        "summary": "Get received emails",
        "description": "Retrieve all received emails for the authenticated user with filtering, search, and pagination options.",
        "tags": [
          "Mail"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "number"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "processed",
                "failed"
              ]
            },
            "required": false
          },
          {
            "in": "query",
            "name": "domain",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "timeRange",
            "schema": {
              "type": "string",
              "enum": [
                "24h",
                "7d",
                "30d",
                "90d"
              ]
            },
            "required": false
          },
          {
            "in": "query",
            "name": "includeArchived",
            "schema": {
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "emailAddress",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "emailId",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of received emails with pagination and filtering options",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "emails": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "emailId": {
                            "type": "string"
                          },
                          "messageId": {
                            "type": "string",
                            "nullable": true
                          },
                          "subject": {
                            "type": "string"
                          },
                          "from": {
                            "type": "string"
                          },
                          "fromName": {
                            "type": "string",
                            "nullable": true
                          },
                          "recipient": {
                            "type": "string"
                          },
                          "preview": {
                            "type": "string"
                          },
                          "receivedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "isRead": {
                            "type": "boolean"
                          },
                          "readAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "isArchived": {
                            "type": "boolean"
                          },
                          "archivedAt": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "hasAttachments": {
                            "type": "boolean"
                          },
                          "attachmentCount": {
                            "type": "number"
                          },
                          "parseSuccess": {
                            "type": "boolean",
                            "nullable": true
                          },
                          "parseError": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "limit": {
                          "type": "number"
                        },
                        "offset": {
                          "type": "number"
                        },
                        "hasMore": {
                          "type": "boolean"
                        }
                      }
                    },
                    "filters": {
                      "type": "object",
                      "properties": {
                        "uniqueDomains": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-v2-mail",
        "summary": "Reply to an email",
        "description": "Create a reply to a received email. This is a convenience endpoint that calls the dedicated reply endpoint internally.",
        "tags": [
          "Mail"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "emailId": {
                    "type": "string",
                    "nullable": false
                  },
                  "to": {
                    "type": "string",
                    "nullable": false
                  },
                  "subject": {
                    "type": "string",
                    "nullable": false
                  },
                  "textBody": {
                    "type": "string",
                    "nullable": true
                  },
                  "htmlBody": {
                    "type": "string",
                    "nullable": true
                  },
                  "attachments": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "filename": {
                          "type": "string"
                        },
                        "contentType": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        }
                      }
                    },
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reply sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "originalEmailId": {
                      "type": "string"
                    },
                    "replyData": {
                      "type": "object",
                      "properties": {
                        "to": {
                          "type": "string"
                        },
                        "subject": {
                          "type": "string"
                        },
                        "hasTextBody": {
                          "type": "boolean"
                        },
                        "hasHtmlBody": {
                          "type": "boolean"
                        },
                        "attachmentCount": {
                          "type": "number"
                        }
                      }
                    },
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}